Introduce RawReservation
