+++ robozonky-api/src/main/java/com/github/robozonky/api/remote/entities/MyReservation.java
@@ -1,0 +1,52 @@
+/*
+ * Copyright 2019 The RoboZonky Project
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package com.github.robozonky.api.remote.entities;
+
+import java.time.OffsetDateTime;
+import javax.xml.bind.annotation.XmlElement;
+
+public class MyReservation extends BaseEntity {
+
+    private long id;
+    private OffsetDateTime timeCreated;
+    private OffsetDateTime deadline;
+    private int reservedAmount;
+
+    MyReservation() {
+        // for JAXB
+    }
+
+    @XmlElement
+    public long getId() {
+        return id;
+    }
+
+    @XmlElement
+    public OffsetDateTime getTimeCreated() {
+        return timeCreated;
+    }
+
+    @XmlElement
+    public OffsetDateTime getDeadline() {
+        return deadline;
+    }
+
+    @XmlElement
+    public int getReservedAmount() {
+        return reservedAmount;
+    }
+}
